Use "slim" image for `pure-rust-build`
This changes `bookworm` to `bookworm-slim`. Information about the
Debian "slim" images is available in the "Image Variants" section
of the "About" page at https://hub.docker.com/_/debian. Which files
are removed changes over time, but I don't think files not present
in the non-slim image will be added to the slim image. Even if
something like that were to happen, the check in the previous
commit should identify an environment that has become unsuitable
for this job.

Using the smaller image should make the job a bit faster, since
that image should be a bit faster to download.