Remove the old disabled Dependabot configuration
In the past (#143), Dependabot version updates had been used to
keep `cargo` dependencies up to date. This was removed in favor of
doing manual updates based on automatic reports from `cargo deny`
and the old `dependabot.yml` was kept but renamed to disable it and
point people to #144 to learn about the change. Since then,
Dependabot security updates, which are distinct from Dependabot
version updates, were enabled (see #1254), and later, Dependabot
version updates were reintroduced for GitHub Actions only (#1357).
At that point, there were two Dependabot-related YAML files: the
old disabled one, and the new one for GHA.

This removes the old one, explaining the situation in a comment in
the new one, including a link to #144.

While doing so, this also adjusts the YAML code style there, to
bring it in line with the style of most other YAML files in the
repository.