Have `wasm` job fail the workflow but not block auto-merge
The WebAssembly CI job definition had an unconditional job-level
`continue-on-error` set to `true`, so that it would always report a
successful conclusion, even on failure. This made sense before PR
auto-merge was set up. But these jobs do not typically fail, and
their failures should be more apparent, even if they should not
yet block PRs from being auto-merged.

This commit changes the `wasm` jobs so they are able to fail and
cause the workflow as a whole to have a failing status, while still
not blocking PR auto-merge. `continue-on-error` is removed from the
job definition, but it is also no longer made a dependency of the
required `tests-pass` check.