completion: improve docs for activation
It was not clear if users should activate both sets of completions if they
wanted the dynamic ones.

The shell commands for activating the dynamic completions are aligned with the
static version. Users can choose / are responsible themselves to add the
activation to their shell config if they want them in every shell instance.

For fish specifically, `script | source` is more idiomatic than
`source (script | psub)`.