Update a journey test `tree` use delcaration to `find`
This also stops installing `tree` on CI.

The `tree` package provides the `tree` command, which is useful but
does not currently appear to be used anywhere. At one time, various
journey tests had used it to generate output for comparison, and
this usage/dependency was declared with `with_program` in 688280b,
which was also when the first command to install it was added to
CI.

Later, `tree` was found to be unportable or otherwise unsuitable in
the counts it output, and all uses of it were replaced with `find`
in 3c1bfd5 (#861). At that point, most `with_program` declarations
for it were changed to `find`, but one of them in the `ein` journey
tests was not replaced at that time.

That test does use `find` (and not `tree`), and should declare its
use of `find` rather than `tree`. This commit makes that change.
One benefit of doing so is that it is possible to run the tests on
a CI (or other) system without `tree` without skipping any tests,
so this also removes it from the lists of packages to install with
`apt-get` and `brew` in CI jobs.