Configure LF line endings for all shell scripts
This adjustes `.gitattributes` to normalize line endings to LF
(Unix style line endings) for all shell scripts in the project, not
just shell scripts that are fixture scripts. Previously, shell
scripts related to journey tests were not normalized this way.

This is simpler to write, and `bash` and POSIX shell scripts are
required to have LF line endings to be correct in general. On
Windows, some `bash` builds (including the MSYS2 `bash` provided as
part of the Git Bash environment) are patched to automatically
accept Windows-style line endings, but I believe they are still
considered technically incorrect. In practice, the effect of not
using LF line endings when working with shell scripts on Windows
may usually be where other tools are involved: ShellCheck, and at
least some language servers for `bash` script editing, complain if
CRLF line endings are used.