Fix gix-url baseline script instead of taking ownership
Taking ownership of cloned files in the container only fixed one
failing test, `gix-url::baseline run`. This had been failing with a
git `safe.directory` error in the the gix-url `make_baseline.sh`
fixture script. That failure was also reproduced locally by
recursively `chown`ing the cloned files to another user (while
preserving write permissions for the first user via the group).

Because the tests shouldn't unnecessarily depend on starting out in
a repository (nor anything about the outer repository they start
in), this should probably be considered a test bug that is not
specific to the current container setup on CI that triggered it.

This commit undoes the `chown -R` step in the CI workflow that had
previously worked around this, and instead fixes it by having the
gix-url `make_baseline.sh` fixture script run its `git fetch-pack`
commands in a temporary repo created in a subdirectory of the
fixture directory (and deleted afterwards).