Check for unrecognized `*-sys` dependencies
This adds another step to `pure-rust-build` that fails -- and fails
the job -- when there are any dependencies named `*-sys` other than
`linux-raw-sys`, which is known about.

(This is independent of the use of C in `ring` -- discussed
in #1681, #1682, and #1684 -- because `ring` is not, and does not
use, a `*-sys` dependency.)

This should fail prior to 3506afb (#1684) and pass afterwards.