Also do the title check if the release is named for testing
Now that the case where the release's name does not cause it to be
treated as a test release, but where the release is not announced
to a non-test discussion thread because it was found not to be
published (or was published but was marked back to draft), has been
manually tested, this commit uncomments the name-check case.

That is to say that this makes it so a release named as being for
testing, rather than being a "real" release, will trigger the
discussion title check. That way, releases that exist just to test
releasing are not inadvertently announced to discussion threads
intended for notifying about actual releases.

This check is done first, since it is the primary case for not
announcing a release that is otherwise eligible to be announced,
and because the log message for it makes more sense when both
reasons to do the title check apply.

At least for now, when the release name indicates the release is
for testing, the draft status of the release is not retrieved or
examined.

(However, as a possible future direction, it might be
worthwhile to check all three things -- release name, release
draft status, and whether the discussion is titled indicating it
is for testing -- and report them in the GitHub Actions
workflow/job log, for all combinations.)

This commit also adds a comment on the `announce-release` job
saying what it's for, as we have on the other `*-release` jobs.