Manually create tag if 32-bit `git` rejects it
This is a workaround for the problem where `git tag` rejects the
timestamp `4147483646` from `generation_number_overflow.sh` as out
of range, even though it has already been accepted to create a
commit and `git tag` is obtaining it from the commit metadata.

The test already passed if run without `GIX_TEST_IGNORE_ARCHIVES`,
even though the commit and tag metadata in the committed archive
repo has this timestamp interpreted as from 1970 rather than in the
far future. The change here just allows the fixture script to
create the repository on a 32-bit system where `git` would refuse
to make the tag, by manually creating a loose tag. (Creating a
packed tag would also work, but would not improve the ambiguity.)