Use `it check-mode` instead of `check-mode.sh`; test on CI
- Delete the old `etc/check-mode.sh` script.
- Update `justfile` recipe `check-mode` to use internal-tools.
- Modify and expand `check-mode` CI job to use internal-tools.
- Temporarily make `check-mode` job a matrix job to test platforms.

The latter of those changes should be undone once things looks like
they are working. I have manually tested `it check-mode` on Arch
Linux and Windows in the same ways `check-mode.sh` was tested (and
unlike with a shell-script based approach, it is not anticipated to
differ on macOS). So trying three platforms on CI is a secondary
supporting strategy to catch possible problems.