Run `check-mode` recipe on CI via `ci-test` recipe
This adds the `check-mode` recipe to the `test` recipe and, more
significantly because of its effect on CI, to the `ci-test` recipe.

This causes the CI `test` job to run the `check-mode` recipe (and
thus `it check-mode` through it).

While the separate `check-mode` CI job was useful for developing
the tool, it imposes extra work, due to the need to build
dependencies that are, to a substantial extent, shared with those
already present (and cached) for the `test` job. Now that it is
run via the `test` CI job, the separate `check-mode` CI job is
removed.