feat: add `topo::Builder::with_tips` fn for adding tips after from_iters
Currently, `Builder::from_iters` takes as arguments both the tips and
ends. Previously, this would be the only possibility for specifying
either of them during the building process. To enhance the
builder-likeness of `Builder`, we recently introduced a fn for adding
additional ends.

This change introduces a fn which allows adding additional tips after
initial construction of a "fresh" `Builder`, allowing for more usage
patterns.