feat: add `topo::Builder::new` fn for creating a bare builder
Previously, the only way to create a `Builder` would be via
`Builder::from_iters`. That fn takes as arguments both the tips and
ends, and used to be the only possibility for specifying either of them
during the building process. However, in order to enhance the builder-
likeness of `Builder`, we recently introduced fns `with_tips` and
`with_ends` for adding additional tips and ends.

With those fns, the only component which needs to be supplied up-front
is `find`. Users can specify and empty `Iterator` and `None` for `tips`
and `ends` respectively when calling `Builder::from_iters` and add
additional tips and ends at their leisure, without the need to chain
`Iterator`s or collecting them in some external data structure.

Now, calling `Builder::from_iters` with a empty lists for tips and ends
is a bit awkward. Thus, we provide a new method for creating a bare
`Builder`.